requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
dask==2024.6.0
Django==5.0.6
findspark==2.0.1
frozendict==2.4.4
fsspec==2024.6.0
html5lib==1.1
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
joblib==1.4.2
locket==1.0.0
lxml==5.2.2
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
partd==1.4.2
peewee==3.17.5
platformdirs==4.2.2
py4j==0.10.9.7
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sklearn==0.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.40
zipp==3.19.2
